translate: fix MapLiteral in range-over-literal
Using composite literals in 'for' or 'if' statements is tricky in Go.
The parser will interpret a '{' as beginning the body of the 'for' or
'if' unless it knows for sure that the '{' begins a composite literal.

Concretely, 'for range map[int]int{1:1} {}' parses as you expect it,
but 'for range MapLiteral[K, V]{{K: 1, V: 1}} {}' does not. (Arguably
it could parse correctly because today the '[K, V]' indexing implies
that 'MapLiteral' is a type and not a value that is indexed... but
that is as an aside.)

Sidestep all literal problems by translating map literals into function
calls. Syntax does not really matter because this function is only
called from generated code.

Also fix sorting in translate testdata.