simulation: pass syscall always for nicer logs
Make stdout and stderr writes have the right machine and goroutine.
Source is still funny... but maybe that is okay.

Always pass around a syscallabi.Syscall to all syscall handlers. This
means they can refer to the calling goroutine, and later perhaps to a
gosimruntime.Step that refers to the calling moment.