simulation: log syscall info on calling goroutine
Add tooling for backtracing other goroutines. Modify logging in syscall
handlers to log as-if the log came from the calling goroutine.

Syscall logs are much more pleasant to look at, but I worry that logging
will be confusing if the stack do not actually match up. Left some
TODOs to think about that.